+++ src/main/java/com/coobird/staticlogistics/api/type/ToolMode.java
@@ -9,8 +9,7 @@ public enum ToolMode implements StringRepresentable {
     LINK_AS_INSERT("link_as_input", ChatFormatting.AQUA, 1),
     LINK_AS_EXTRACT("link_as_output", ChatFormatting.GOLD, 2),
     REMOVE("remove", ChatFormatting.RED, 3),
-    NODE_CONFIG("node_config", ChatFormatting.LIGHT_PURPLE, 4),
-    CONTAINER_CONFIG("container_config", ChatFormatting.GREEN, 5);
+    NODE_CONFIG("node_config", ChatFormatting.LIGHT_PURPLE, 4);
 
     private final String name;
     private final ChatFormatting color;